Refactor: moves Unique Process Numbering validation logic to UniqueProcessNumberingFactory.
