+++ application/src/main/java/io/github/ricardosander/jusbrasilprocessservice/application/RetrieveProcessGateway.java
@@ -1,7 +1,9 @@
 package io.github.ricardosander.jusbrasilprocessservice.application;
 
+import io.github.ricardosander.jusbrasilprocessservice.domain.UniqueProcessNumbering;
+
 import java.util.Optional;
 
 interface RetrieveProcessGateway {
-    Optional<Process> execute(String uniqueProcessNumbering);
+    Optional<Process> execute(UniqueProcessNumbering uniqueProcessNumbering);
 }
